+++ src/model/EnergyDrinkModel.java
@@ -6,7 +6,7 @@ import org.jbox2d.dynamics.World;
 
 public class EnergyDrinkModel extends AbstractPowerUpModel {
 	
-	private final float HP_BOOST = 0.5f;
+	private final double HP_BOOST = 0.5;
 
 	public EnergyDrinkModel(World w, Vec2 pixelPos, int id) {
 		super(w, pixelPos, id);
